+++ grid_map_filters/include/grid_map_filters/SlidingWindowMathExpressionFilter.hpp
@@ -17,7 +17,7 @@
 
 #include <string>
 
-#include "EigenLab/EigenLab.h"
+#include "EigenLab/EigenLab.hpp"
 
 namespace grid_map
 {
